Add -b, --bugs: Also show status of `bugs' file
Add support for the flags shown. If given, these flags show a cyan
'B' in the output for each repository entry that contains a file
titled `bugs' in it's root directory. If a repository contains no
other changes, it is still shown in the output. This feature works
nicely with the -s flag. This feature was meant to be paired with
the `b' tool in the AmateurECE/Tools repository.