Add -r,--remotes to check remote branches
This flag checks the status of the remote refs to determine if there
are any discrepancies (e.g. if a pull is needed, or if local has
diverged). It slows the program down greatly because it must run
`git remote update` on all repositories, which talks to the remote
repository over the network.